endpoint with user recipes
